Update viewer so registering with the pyomo command doesn't take 14 seconds
